refactor: phase A/B contracts normalization and primitives commons path
