+++ model/geo_primitives/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 geo_foundation = { path = "../geo_foundation" }
 geo_contracts = { path = "../geo_contracts" }
 geo_core = { path = "../geo_core" }
+geo_commons = { path = "../geo_commons" }
 analysis = { path = "../../foundation/analysis" }
 
 [features]
